refactor: relocate useAttrsOnly useComponentId useListeners
