+++ client-app/core/composables/index.ts
@@ -1,9 +1,7 @@
 export * from "./useAppContext";
-export * from "./useAttrsOnly";
 export * from "./useBreadcrumbs";
 export * from "./useCartValidationErrorTranslator";
 export * from "./useCategoriesRoutes";
-export * from "./useComponentId";
 export * from "./useCountries";
 export * from "./useCurrency";
 export * from "./useErrorsTranslator";
@@ -11,7 +9,6 @@ export * from "./useFetch";
 export * from "./useGoogleAnalytics";
 export * from "./useIdentityErrorTranslator";
 export * from "./useLanguages";
-export * from "./useListeners";
 export * from "./useNavigations";
 export * from "./usePageHead";
 export * from "./usePagesWithFullCartLoad";
