+++ client-app/ui-kit/composables/index.ts
@@ -1,0 +1,3 @@
+export * from "./useAttrsOnly";
+export * from "./useComponentId";
+export * from "./useListeners";
