+++ client-app/ui-kit/composables/useAttrsOnly.ts
